+++ src/main/index.ts
@@ -59,6 +59,7 @@ import { TrackpadGestureLedger } from './trackpad-gesture'
 import { registerConverterIpc } from '../core/converter/register-ipc'
 import { registerNodeDependencyIpc } from '../core/node-dependencies/register-ipc'
 import { registerOllamaIpc } from '../core/ollama/register-ipc'
+import { registerUniGetUiIpc } from '../core/unigetui/register-ipc'
 import { registerOpenWebUiHosting } from './open-webui-hosting'
 import { registerMinecraftIpc } from '../core/minecraft/register-ipc'
 import { registerAwsIdentityIpc } from '../core/aws-identity'
@@ -2278,6 +2279,7 @@ app.whenReady().then(async () => {
   const nodeDependencyService = registerNodeDependencyIpc(corePlatform)
   const awsWizardModels = new AwsWizardModelService(nodeDependencyService)
   registerOllamaIpc(corePlatform)
+  registerUniGetUiIpc(corePlatform)
   registerOpenWebUiHosting(getMainWindow, app.getPath('userData'))
   minecraftServers = registerMinecraftIpc(corePlatform).manager
   registerAwsIdentityIpc(corePlatform, {