+++ src/server/handlers/index.ts
@@ -11,6 +11,7 @@ import { registerNodeDependencyIpc } from '../../core/node-dependencies/register
 import { registerAwsResourceIpc } from '../../core/aws-resource-register-ipc'
 import { AwsWizardModelService } from '../../core/aws-wizard/service'
 import { registerOllamaIpc } from '../../core/ollama/register-ipc'
+import { registerUniGetUiIpc } from '../../core/unigetui/register-ipc'
 import { registerMinecraftIpc } from '../../core/minecraft/register-ipc'
 import { registerTorrentIpc } from '../../core/torrent/register-ipc'
 import { registerVirtualMachineIpc } from '../../core/virtual-machine/register-ipc'
@@ -91,6 +92,7 @@ export function registerCoreHandlers(
   const nodeDependencyService = registerNodeDependencyIpc(platform)
   const awsWizardModels = new AwsWizardModelService(nodeDependencyService)
   registerOllamaIpc(platform)
+  registerUniGetUiIpc(platform)
   const { manager: minecraftServers } = registerMinecraftIpc(platform)
   registerTorrentIpc(platform)
   const { manager: virtualMachineManager } = registerVirtualMachineIpc(platform)